% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{iQualityR.spc-package}
\alias{iQualityR.spc}
\alias{iQualityR.spc-package}
\title{iQualityR.spc}
\description{
Statistical process control (SPC) charts for the iQualityR ecosystem.
Implements the full Minitab Stat > Control Charts menu (variables, attributes,
time-weighted, multivariate, rare-event charts) plus Nelson rules detection.
}
\author{
\strong{Maintainer}: iQualityR Team \email{iQualityR@example.com}

}
\keyword{internal}
