% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcAnalyzer.R
\name{SpcAnalyzer}
\alias{SpcAnalyzer}
\title{SpcAnalyzer}
\description{
Analyzer for statistical process control charts. Inherits \code{IqrAnalyzerBase}
and dispatches to chart-type specific computation routines. Reuses
\code{iQualityR.stat} for control-limit calculation, sigma estimation, and
Nelson rules detection where possible.
}
\section{Super class}{
\code{\link[iQualityR.core:IqrAnalyzerBase]{iQualityR.core::IqrAnalyzerBase}} -> \code{SpcAnalyzer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SpcAnalyzer-run}{\code{SpcAnalyzer$run()}}
\item \href{#method-SpcAnalyzer-to_spc_result}{\code{SpcAnalyzer$to_spc_result()}}
\item \href{#method-SpcAnalyzer-clone}{\code{SpcAnalyzer$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="get_results"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-get_results'><code>iQualityR.core::IqrAnalyzerBase$get_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="initialize"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-initialize'><code>iQualityR.core::IqrAnalyzerBase$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="reset"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-reset'><code>iQualityR.core::IqrAnalyzerBase$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="set_datatable"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-set_datatable'><code>iQualityR.core::IqrAnalyzerBase$set_datatable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="set_diagnostic"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-set_diagnostic'><code>iQualityR.core::IqrAnalyzerBase$set_diagnostic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="set_raw_output"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-set_raw_output'><code>iQualityR.core::IqrAnalyzerBase$set_raw_output()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="set_statistic"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-set_statistic'><code>iQualityR.core::IqrAnalyzerBase$set_statistic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrAnalyzerBase" data-id="setup"><a href='../../iQualityR.core/html/IqrAnalyzerBase.html#method-IqrAnalyzerBase-setup'><code>iQualityR.core::IqrAnalyzerBase$setup()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpcAnalyzer-run"></a>}}
\if{latex}{\out{\hypertarget{method-SpcAnalyzer-run}{}}}
\subsection{Method \code{run()}}{
Run SPC analysis. Overrides base \code{run()} to accept the
SPC-specific signature.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpcAnalyzer$run(
  x = NULL,
  subgroup = NULL,
  count = NULL,
  sample_size = NULL,
  plan,
  data = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Numeric vector of measurements (or subgroup means).}

\item{\code{subgroup}}{Optional subgroup vector.}

\item{\code{count}}{Optional defect/defective count vector (attributes charts).}

\item{\code{sample_size}}{Optional sample-size vector (attributes charts).}

\item{\code{plan}}{SpcPlan object.}

\item{\code{data}}{Optional data.frame for multivariate charts.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpcAnalyzer-to_spc_result"></a>}}
\if{latex}{\out{\hypertarget{method-SpcAnalyzer-to_spc_result}{}}}
\subsection{Method \code{to_spc_result()}}{
Get a compact result list suitable for the run_*() contract.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpcAnalyzer$to_spc_result()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpcAnalyzer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SpcAnalyzer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpcAnalyzer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
